+++ src/main/kotlin/app/revanced/cli/command/ListCompatibleVersions.kt
@@ -1,9 +1,9 @@
 package app.revanced.cli.command
 
 import app.revanced.library.PackageName
-import app.revanced.library.PatchUtils
 import app.revanced.library.VersionMap
-import app.revanced.patcher.PatchBundleLoader
+import app.revanced.library.mostCommonCompatibleVersions
+import app.revanced.patcher.patch.loadPatchesFromJar
 import picocli.CommandLine
 import java.io.File
 import java.util.logging.Logger
@@ -12,17 +12,17 @@ import java.util.logging.Logger
     name = "list-versions",
     description = [
         "List the most common compatible versions of apps that are compatible " +
-            "with the patches in the supplied patch bundles.",
+            "with the patches from RVP files.",
     ],
 )
 internal class ListCompatibleVersions : Runnable {
-    private val logger = Logger.getLogger(ListCompatibleVersions::class.java.name)
+    private val logger = Logger.getLogger(this::class.java.name)
 
     @CommandLine.Parameters(
-        description = ["Paths to patch bundles."],
+        description = ["Paths to RVP files."],
         arity = "1..*",
     )
-    private lateinit var patchBundles: Array<File>
+    private lateinit var patchesFiles: Set<File>
 
     @CommandLine.Option(
         names = ["-f", "--filter-package-names"],
@@ -38,8 +38,6 @@ internal class ListCompatibleVersions : Runnable {
     private var countUnusedPatches: Boolean = false
 
     override fun run() {
-        val patches = PatchBundleLoader.Jar(*patchBundles)
-
         fun VersionMap.buildVersionsString(): String {
             if (isEmpty()) return "Any"
 
@@ -58,8 +56,9 @@ internal class ListCompatibleVersions : Runnable {
                 appendLine(versions.buildVersionsString().prependIndent("\t"))
             }
 
-        PatchUtils.getMostCommonCompatibleVersions(
-            patches,
+        val patches = loadPatchesFromJar(patchesFiles)
+
+        patches.mostCommonCompatibleVersions(
             packageNames,
             countUnusedPatches,
         ).entries.joinToString("\n", transform = ::buildString).let(logger::info)